+++ ServerJars.Demo.Console/Program.cs
@@ -21,12 +21,13 @@ ResetConsoleColor();
 var types = await serverJar.GetTypes();
 Console.WriteLine(JsonSerializer.Serialize(types, jsonOptions));
 
-// GetTypes.ToList() extension
+// GetTypes.AsDictionary() extension
 SetConsoleColor(ConsoleColor.White, ConsoleColor.Red);
-Console.WriteLine("\nAPI call - GetTypes.ToList() extension:\n");
+Console.WriteLine("\nAPI call - GetTypes.AsDictionary() extension:\n");
 ResetConsoleColor();
 
-types.ToList().ForEach(t => Console.WriteLine(t.ToString()));
+var dict = types.AsDictionary();
+Console.WriteLine(string.Join(Environment.NewLine, dict.Select((kv) => $"{kv.Key}: {string.Join(", ", kv.Value)}")));
 
 
 // GetDetails